+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart'; //import dotenv libary für
 import 'login.dart'; // import der login.dart damit das Authgate weiß wo es hinleiten muss
 import 'navbar.dart'; // Import der Datei wo die nav bar also die NAvigationsleite ist damit sie eingebunden werden kann
 import 'user_SessionManager.dart';
+import 'achievment_runner.dart';
 
 Future<void> main() async { //main funktion wird immer als erstes ausgeführt
   WidgetsFlutterBinding.ensureInitialized(); // verbindung zwischen framework flutetr und der engine flutter sozusagen die Brücke zwischen der hardware und der app
@@ -88,12 +89,12 @@ class _MyAppState extends State<MyApp> {
 }
 
 class AuthGate extends StatelessWidget { 
-   // GAst bzw. Localer Acount
+
    final bool isGuest;
    final VoidCallback onContinueAsGuest;
-   // Dark / light mode
+
    final Function(ThemeMode) onThemeChanged;
-   // nav bar asugewählte seite
+
    final int selectedIndex;
    final Function(int) onIndexChanged;
 
@@ -129,7 +130,10 @@ class AuthGate extends StatelessWidget {
           onContinueAsGuest: onContinueAsGuest);
         }
 
-        UserSessionmanager.initialize();
+          Future.microtask(() async {
+            await UserSessionmanager.initialize();
+            await AchievementRunner.run();
+          });
 
         return GoogleBottomBar(onThemeChanged: onThemeChanged, initalIndex: selectedIndex, onIndexChanged: onIndexChanged);